requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
anyio==3.6.2
    # via starlette
azure-functions==1.13.3
    # via -r requirements.in
click==8.1.3
    # via uvicorn
cython==0.29.34
    # via pysam
fastapi==0.95.0
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
idna==3.4
    # via anyio
pydantic==1.10.7
    # via fastapi
pysam==0.21.0
    # via -r requirements.in
sniffio==1.3.0
    # via anyio
starlette==0.26.1
    # via fastapi
typing-extensions==4.5.0
    # via pydantic
uvicorn==0.21.1
    # via -r requirements.in
